Added Filtering feature, redesigned layout of home page, implemented the "Download" feature
I created a button that allows the user to filter between two years and see it live on the table.

I also reformatted the layout of the entire site so that it looks more clean and aesthetic.

Implemented and designed the download buttons that allows the user to get a .csv or .pdf file of the report and data.